Changed file structure of reviews to organize subcomponents
